refactor(model): organise styleTransfer and other files
