+++ model/StyleTransfer/ContentLoss.py
@@ -1,0 +1,19 @@
+import torch.nn as nn
+import torch.nn.functional as F
+
+
+class ContentLoss(nn.Module):
+    def __init__(
+        self,
+        target,
+    ):
+        super(ContentLoss, self).__init__()
+        # we 'detach' the target content from the tree used
+        # to dynamically compute the gradient: this is a stated value,
+        # not a variable. Otherwise the forward method of the criterion
+        # will throw an error.
+        self.target = target.detach()
+
+    def forward(self, input):
+        self.loss = F.mse_loss(input, self.target)
+        return input
